feat: add library API
